backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==1.3.0
dj-email-url==1.0.6
dj-rest-auth==2.1.11
Django==4.0.10
django-allauth==0.48.0
django-cache-url==3.4.4
django-cors-headers==3.10.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
environs==9.5.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
marshmallow==3.19.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==5.3.0
celery[redis]
Pillow==9.5.0
pandas==2.0.1
sendgrid==6.10.0
openai==0.27.7
pan-os-upgrade==1.3.10